pyproject.toml
[project]
name = "llm_router"
version = "0.1.0"
description = "LLM router"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "pyyaml",
    "httpx",
    "python-dotenv",
    "litellm",
    "openai",
    "google-genai",
    "cachetools",
    "pytest-asyncio",
    "pytest-cov"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C90", # mccabe complexity
    "B",   # bugbear
    "UP",  # pyupgrade
    "YTT", # flake8-2020
    "FBT", # flake8-boolean-trap
    "Q",   # flake8-quotes
    "RUF", # ruff-specific rules
]
ignore = ["E501", "B008"]  # allow long lines, and function call in argument defaults

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.isort]
known-first-party = ["app", "tests"]

[tool.mypy]
python_version = "3.10"
strict = true
show_error_codes = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "pydantic.*",
    "fastapi.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--asyncio-mode=auto --cov=app --cov-report=term-missing"
testpaths = ["tests"]